+++ CHANGELOG.md
@@ -2,7 +2,12 @@
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
+## [Unreleased]
+### Changed
+- `JSONPointer`: added KDoc, minor code improvements
+- `JSONPointer`: added `root`
+- `JSONPointer`: added `toURIFragment()` and `fromURIFragment()`
+
 ## [0.1] - 2020-07-13
 ### Added
 - `JSONPointer`: initial version
-